+++ src/hapi/pipelines/database/db_org.py
@@ -1,9 +1,8 @@
-"""Admin1 table."""
+"""Org table."""
 from datetime import datetime
 
 from hdx.database.no_timezone import Base
 from sqlalchemy import (
-    Boolean,
     DateTime,
     ForeignKey,
     Integer,
@@ -12,26 +11,24 @@ from sqlalchemy import (
 )
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from hapi.pipelines.database.dblocation import DBLocation  # noqa: F401
+from hapi.pipelines.database.db_orgtype import DBOrgType  # noqa: F401
 
 
-class DBAdmin1(Base):
-    __tablename__ = "admin1"
+class DBOrg(Base):
+    __tablename__ = "org"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    location_ref: Mapped[int] = mapped_column(
-        ForeignKey("location.id", onupdate="CASCADE", ondelete="CASCADE")
-    )
-    code: Mapped[str] = mapped_column(String(128), nullable=False)
+    hdx_link: Mapped[str] = mapped_column(String(1024), nullable=False)
+    acronym = mapped_column(String(32), nullable=False, index=True)
     name: Mapped[str] = mapped_column(String(512), nullable=False)
-    is_unspecified: Mapped[bool] = mapped_column(
-        Boolean, server_default=text("FALSE")
+    org_type_code: Mapped[str] = mapped_column(
+        ForeignKey("org_type.code", onupdate="CASCADE", ondelete="CASCADE")
     )
     reference_period_start: Mapped[datetime] = mapped_column(
-        DateTime, nullable=False
+        DateTime, nullable=False, index=True
     )
     reference_period_end: Mapped[datetime] = mapped_column(
         DateTime, nullable=True, server_default=text("NULL")
     )
 
-    location = relationship("DBLocation")
+    org_type = relationship("DBOrgType")
